Tune SVR model for surrogate

+++ NSGA-II/evaluate.m
@@ -78,37 +78,44 @@ function [indi, evalTime] = evalIndividual(indi, objfun,index,conf)
 %         LSSSSWC, NWPU
 %    Revision: 1.1  Data: 2011-07-25
 %*************************************************************************
-
-% copy files to new tempory folder to avoid conflicts in parallel computing
-if ~isempty(conf)
-   folderName=copyFiles(index,conf); 
-
-% change working directory to temp folder
-cwd=pwd;
-cd ([cwd,'/',folderName]);
-end
-tStart = tic;
-[y, cons] = objfun( indi.var);
-indi.cons=cons;
-evalTime = toc(tStart);
-
-% Save the objective values and constraint violations
-indi.obj = y;
-if( ~isempty(indi.cons) )
-    idx = find( cons>0 );
-    if( ~isempty(idx) )
-        indi.nViol = length(idx);
-        indi.violSum = sum( abs(cons(idx)) );
-    else
-        indi.nViol = 0;
-        indi.violSum = 0;
+if indi.expensive==0
+    % copy files to new tempory folder to avoid conflicts in parallel computing
+    if ~isempty(conf)
+        folderName=copyFiles(index,conf);
+        
+        % change working directory to temp folder
+        cwd=pwd;
+        cd ([cwd,'/',folderName]);
     end
-end
-
-if ~isempty(conf)
-cd ../
-% delete the temp folder
-rmdir(folderName,'s');
+    tStart = tic;
+    [y, cons] = objfun( indi.var);
+    indi.cons=cons;
+    evalTime = toc(tStart);
+    
+    % Save the objective values and constraint violations
+    indi.obj = y;
+    if( ~isempty(indi.cons) )
+        idx = find( cons>0 );
+        if( ~isempty(idx) )
+            indi.nViol = length(idx);
+            indi.violSum = sum( abs(cons(idx)) );
+        else
+            indi.nViol = 0;
+            indi.violSum = 0;
+        end
+    end
+    
+    if ~isempty(conf)
+        cd ../
+        % delete the temp folder
+        rmdir(folderName,'s');
+    end
+    % change the flag to true
+    indi.expensive=1;
+else
+    evalTime=0;
+    % the flag keep to be true although it's not evaluated in current
+    % generation
 end
 
 function folderName=copyFiles(folderIndex,conf)
